added lab4

+++ src/com/studying/util/FifthLabAlternativesAndCriteria.java
@@ -4,21 +4,6 @@ import java.util.List;
 
 public class FifthLabAlternativesAndCriteria {
     public static final List<List<Integer>> ALTERNATIVES_AND_CRITERIA = List.of(
-//            List.of(7, 10, 3, 1, 5, 10, 3, 8, 9, 7, 6, 5),
-//            List.of(5, 6, 2, 5, 4, 8, 7, 5, 3, 5, 10, 8),
-//            List.of(10, 9, 4, 5, 9, 9, 9, 5, 5, 3, 2, 6),
-//            List.of(7, 4, 8, 6, 7, 2, 8, 4, 9, 2, 6, 5),
-//            List.of(10, 10, 9, 9, 7, 5, 4, 8, 7, 1, 7, 10),
-//            List.of(3, 9, 9, 8, 4, 7, 4, 4, 3, 6, 8, 8),
-//            List.of(10, 5, 9, 10, 7, 8, 5, 10, 5, 6, 1, 9),
-//            List.of(9, 3, 4, 9, 6, 4, 1, 5, 5, 5, 4, 2),
-//            List.of(3, 2, 6, 5, 8, 8, 2, 8, 10, 5, 5, 10),
-//            List.of(2, 2, 2, 7, 2, 1, 8, 5, 10, 9, 2, 6),
-//            List.of(3, 2, 5, 4, 7, 1, 5, 9, 5, 9, 9, 10),
-//            List.of(5, 2, 1, 8, 7, 4, 7, 7, 10, 6, 3, 8),
-//            List.of(8, 8, 8, 3, 7, 3, 1, 4, 10, 3, 7, 10),
-//            List.of(6, 10, 2, 10, 9, 1, 10, 9, 6, 9, 7, 3),
-//            List.of(6, 7, 10, 10, 7, 6, 3, 3, 7, 9, 10, 6)
             List.of(6, 1, 9, 8, 3, 5, 9, 5, 7, 2, 1, 5),
             List.of(6, 9, 1, 6, 1, 10, 1, 6, 7, 4, 7, 2),
             List.of(3, 3, 10, 3, 3, 2, 3, 10, 7, 7, 2, 8),
@@ -36,6 +21,5 @@ public class FifthLabAlternativesAndCriteria {
             List.of(2, 10, 5, 5, 8, 5, 7, 6, 8, 6, 3, 9)
     );
 
-    public static final List<Integer> WEIGHTS = List.of(10, 3, 1, 10, 2, 4, 9, 10, 4, 6, 1, 8);
-//        public static final List<Integer> WEIGHTS = List.of(3, 7, 6, 2, 1, 1, 6, 3, 9, 2, 1, 1);
+    public static final List<Double> WEIGHTS = List.of(10d, 3d, 1d, 10d, 2d, 4d, 9d, 10d, 4d, 6d, 1d, 8d);
 }
